ProductRangeSpace typed

+++ src/Multivariate/TensorSpace.jl
@@ -74,7 +74,7 @@ domain(f::ProductSpace)=domain(f.spacesx[1])*domain(f.spacesy)
 Base.getindex(d::ProductSpace,k::Integer)=k==1?d.spacesx:d.spacey
 
 
-space(d::AbstractProductSpace,k::Integer)=d[k]
+space(d::AbstractProductSpace,k)=d[k]
 
 
 for TT in (:ProductDomain,:TensorSpace)
